+++ src/deploy/functions/release/fabricator.ts
@@ -25,7 +25,7 @@ import * as scheduler from "../../../gcp/cloudscheduler";
 import * as utils from "../../../utils";
 import * as services from "../services";
 import { AUTH_BLOCKING_EVENTS } from "../../../functions/events/v1";
-import { getDefaultComputeServiceAgent } from "../checkIam";
+import * as gcb from "../../../gcp/cloudbuild";
 import { getHumanFriendlyPlatformName } from "../functionsDeployHelper";
 
 // TODO: Tune this for better performance.
@@ -436,7 +436,7 @@ export class Fabricator {
     } else if (backend.isScheduleTriggered(endpoint)) {
       const invoker = endpoint.serviceAccount
         ? [endpoint.serviceAccount]
-        : [getDefaultComputeServiceAgent(this.projectNumber)];
+        : [gcb.getDefaultComputeEngineServiceAgent(this.projectNumber)];
       await this.executor
         .run(() => run.setInvokerCreate(endpoint.project, serviceName, invoker))
         .catch(rethrowAs(endpoint, "set invoker"));
@@ -543,7 +543,7 @@ export class Fabricator {
     } else if (backend.isScheduleTriggered(endpoint)) {
       invoker = endpoint.serviceAccount
         ? [endpoint.serviceAccount]
-        : [getDefaultComputeServiceAgent(this.projectNumber)];
+        : [gcb.getDefaultComputeEngineServiceAgent(this.projectNumber)];
     }
 
     if (invoker) {
