+++ src/gcp/cloudbuild.ts
@@ -223,3 +223,13 @@ export async function deleteRepository(
 export function serviceAgentEmail(projectNumber: string): string {
   return `service-${projectNumber}@gcp-sa-cloudbuild.iam.gserviceaccount.com`;
 }
+
+/** Returns the default cloud build service agent */
+export function getDefaultCloudBuildServiceAgent(projectNumber: string): string {
+  return `${projectNumber}@cloudbuild.gserviceaccount.com`;
+}
+
+/** Returns the default compute engine service agent */
+export function getDefaultComputeEngineServiceAgent(projectNumber: string): string {
+  return `${projectNumber}-compute@developer.gserviceaccount.com`;
+}
